Commons Update 0.0.27
This commit upgrades the plugin to commons 0.0.27. Along with this extended applink
and directory resources are implemented as well.

+++ src/test/java/de/aservo/confapi/jira/model/util/MailServerPopBeanUtilTest.java
@@ -23,7 +23,7 @@ public class MailServerPopBeanUtilTest {
         assertEquals(server.getMailProtocol().getProtocol(), bean.getProtocol());
         assertEquals(server.getHostname(), bean.getHost());
         assertEquals(Integer.valueOf(server.getPort()), bean.getPort());
-        assertEquals(server.getTimeout(), bean.getTimeout());
+        assertEquals(server.getTimeout(), (long) bean.getTimeout());
         assertEquals(server.getUsername(), bean.getUsername());
         assertNull(bean.getPassword());
     }
